x/xarchain/keeper/query_cblock.go
package keeper

import (
	"context"

	"xarchain/x/xarchain/types"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) Cblock(goCtx context.Context, req *types.QueryGetCblockRequest) (*types.QueryGetCblockResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(goCtx)

	val, found := k.GetCblock(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetCblockResponse{Cblock: val}, nil
}
